feat: multiplatform paging
