+++ gradle/libs.versions.toml
@@ -30,7 +30,7 @@ kover = "0.7.1"
 multiplatform-settings = "1.1.1"
 # if you update sqlDelight check if https://github.com/cashapp/sqldelight/issues/4154 is fixed
 # and delete the work around in the dev.mk file
-sqldelight = "2.0.0"
+sqldelight = "2.0.1"
 sqlcipher-android = "4.5.5"
 pbandk = "0.14.2"
 turbine = "1.0.0"
