--- persistence/src/androidMain/kotlin/com/wire/kalium/persistence/dao/message/KaliumOffsetQueryPagingSource.kt
@@ -1,116 +1,0 @@
-/*
- * Wire
- * Copyright (C) 2023 Wire Swiss GmbH
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see http://www.gnu.org/licenses/.
- */
-package com.wire.kalium.persistence.dao.message
-
-import androidx.paging.PagingSource
-import androidx.paging.PagingState
-import app.cash.paging.PagingSourceLoadParams
-import app.cash.paging.PagingSourceLoadParamsAppend
-import app.cash.paging.PagingSourceLoadParamsPrepend
-import app.cash.paging.PagingSourceLoadParamsRefresh
-import app.cash.paging.PagingSourceLoadResult
-import app.cash.paging.PagingSourceLoadResultInvalid
-import app.cash.paging.PagingSourceLoadResultPage
-import app.cash.sqldelight.Query
-import app.cash.sqldelight.SuspendingTransacter
-import app.cash.sqldelight.Transacter
-import app.cash.sqldelight.TransacterBase
-import app.cash.sqldelight.TransactionCallbacks
-import app.cash.sqldelight.db.QueryResult
-import app.cash.sqldelight.db.SqlCursor
-import kotlinx.coroutines.withContext
-import kotlin.coroutines.CoroutineContext
-import kotlin.properties.Delegates
-
-/**
- * This class is duplicated from sqlDelight []OffsetQueryPagingSource] so we can pass an initial offset instead of depending
- * only on the offset result coming from the queryProvider callback.
- */
-@Suppress("ComplexMethod")
-internal class KaliumOffsetQueryPagingSource<RowType : Any>(
-    private val queryProvider: (limit: Int, offset: Int) -> Query<RowType>,
-    private val countQuery: Query<Int>,
-    private val transacter: TransacterBase,
-    private val initialOffset: Int,
-    private val context: CoroutineContext,
-) : WireQueryPagingSource<Int, RowType>() {
-    override val jumpingSupported get() = true
-
-    override suspend fun load(
-        params: PagingSourceLoadParams<Int>,
-    ): PagingSourceLoadResult<Int, RowType> = withContext(context) {
-        val key = params.key ?: initialOffset
-        val limit = when (params) {
-            is PagingSourceLoadParamsPrepend<*> -> minOf(key, params.loadSize)
-            else -> params.loadSize
-        }
-        val getPagingSourceLoadResult: TransactionCallbacks.() -> PagingSourceLoadResultPage<Int, RowType> = {
-            val count = countQuery.executeAsOne()
-            val offset = when (params) {
-                is PagingSourceLoadParamsPrepend<*> -> maxOf(0, key - params.loadSize)
-                is PagingSourceLoadParamsAppend<*> -> key
-                is PagingSourceLoadParamsRefresh<*> -> if (key >= count) maxOf(0, count - params.loadSize) else key
-                else -> error("Unknown PagingSourceLoadParams ${params::class}")
-            }
-            val data = queryProvider(limit, offset)
-                .also { currentQuery = it }
-                .executeAsList()
-            val nextPosToLoad = offset + data.size
-            PagingSourceLoadResultPage(
-                data = data,
-                prevKey = offset.takeIf { it > 0 && data.isNotEmpty() },
-                nextKey = nextPosToLoad.takeIf { data.isNotEmpty() && data.size >= limit && it < count },
-                itemsBefore = offset,
-                itemsAfter = maxOf(0, count - nextPosToLoad),
-            )
-        }
-        val loadResult = when (transacter) {
-            is Transacter -> transacter.transactionWithResult(bodyWithReturn = getPagingSourceLoadResult)
-            is SuspendingTransacter -> transacter.transactionWithResult(bodyWithReturn = getPagingSourceLoadResult)
-        }
-        (if (invalid) PagingSourceLoadResultInvalid<Int, RowType>() else loadResult) as PagingSourceLoadResult<Int, RowType>
-    }
-
-    override fun getRefreshKey(state: PagingState<Int, RowType>) =
-        state.anchorPosition?.let { maxOf(0, it - (state.config.initialLoadSize / 2)) }
-}
-
-internal abstract class WireQueryPagingSource<Key : Any, RowType : Any> :
-    PagingSource<Key, RowType>(),
-    Query.Listener {
-    protected var currentQuery: Query<RowType>? by Delegates.observable(null) { _, old, new ->
-        old?.removeListener(this)
-        new?.addListener(this)
-    }
-
-    init {
-        registerInvalidatedCallback {
-            currentQuery?.removeListener(this)
-            currentQuery = null
-        }
-    }
-
-    final override fun queryResultsChanged() = invalidate()
-}
-
-internal fun Query<Long>.toInt(): Query<Int> =
-    object : Query<Int>({ cursor -> mapper(cursor).toInt() }) {
-        override fun <R> execute(mapper: (SqlCursor) -> QueryResult<R>) = this@toInt.execute(mapper)
-        override fun addListener(listener: Listener) = this@toInt.addListener(listener)
-        override fun removeListener(listener: Listener) = this@toInt.removeListener(listener)
-    }
